pyproject.toml
[tool.poetry]
name = "telegrambot-price"
version = "0.1.0"
description = ""
authors = ["sajad <fallahdoostsajad@gmal.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
